AdminQ DOL support (#6366)
* Reserve RDMA resources to be created using AdminQ

PD 1 is now reserved for RDMA AdminQ operations and all resources
in the PD are to be created using AdminQ (or dev cmd in the actual
test).


* Add Dev Cmd interface to create RDMA resources

RDMA resources belonging to the Admin PD have to be created
either using the AdminQ or Dev Cmd when the actual test runs
on a need basis. Adding interfaces to help create these resources
on a PD level using Dev Cmds. The only Dev Cmds to be supported
in the final version is to create CQ and AQ. Everything else must
ideally be done using AdminQ.


* Create RDMA resources needed by AdminQ before test phase

RDMA resources needed for AdminQ (CQ, EQ) have to be created first
using dev cmd before AdminQ is created. Added helper functions in
RDMA PD object to access CQ and QP objects using their IDs.


* Add RDMA AdminQ Descriptor formats


* RDMA AdminQ DOL utility functions

Some utility functions to test RDMA AdminQ functionality.


* RDMA AdminQ NOP testcase

Please use the NOP testcase template to add more AdminQ tests.


* Register RDMA PDs with LIF


* Reorder RDMA DOL objects

- CQ moved to endpoint object
- AQ, EQ and AQ_CQ moved to Lif object
- QPs belong to PDs. Every QP gets a CQ


* All device PT resources handled by DOL

As the driver is moving to managing all device table resources,
we make the move in the dev cmds and include device table management
layer in DOL. PT table is now shared between PT entries for RDMA
resources and RSQ/RRQ queues for RC QPs. This makes way for efficient
reuse of resources and avoids preallocating a very large chunk of memory
upfront for these queues to be managed by the admin cmd interface.


* Move RDMA DOL CreateCQ operation to AdminQ

With this checkin RDMA create_cq operation in DOL is handled by
the AdminQ. The useAdmin flag in cq.spec toggles the HAL vs AdminQ
behavior. The plan is to introduce this switch for all resources
created by AdminQ.


* RDMA DOL to AdminQ: Fix #1

Fixes found when moving RDMA DOL Infra creation to AdminQ.
Convert EQ_ID in create_cq Admin Cmd to Big Endian format.


* Do not configure RDMA resources in a DOL dryrun


* Move RDMA DOL CreateMR operation to AdminQ

With this checkin RDMA create_mr operation in DOL is handled by
the AdminQ. The useAdmin flag in mr.spec toggles the HAL vs AdminQ
behavior.


* Prepare RDMA AdminAPIs to optimize AdminQ DOL operations

Preparing the adminapi code to issue admin wqe cmds in batches
and ring doorbell only once. This will work in model only if busy
bit is removed from adminq. Now, doorbell is rung for every admin
wqe, once the busy bit goes, we can easily switch to batch adminQ
cmds.


* Bump up max_eq count to 8

Always have power of 2 resource counts


* Move Endian definitions to scapy wqe descriptions

Scapy has inherent support to convert values to little endian
format if the wqe field definitions have the "LE" qualifier.
No need to explicitly convert the values to LE format.


* CreateQP, ModifyQP move to Admin Cmd

CreateQP and ModifyQP calls now route to Admin Cmd from
dev cmd. The Perf RC QPs (NIC resident) are still using
the dev cmd path until barmap allocation is added to HAL
code for DOL.


* Move Perf RC QP Creation to AdminQ

Perf RC QPs (QPs in NIC) are created and managed in AdminQ.
A dummy barmap area resembling the one created by Nicmgr is
created in HAL lif_init for DOL.


* RDMA HAL manages it's own resources

To accomodate Storage DOL using Dev Cmds


* Initialize DCQCN CB from DOL

Init DCQCN CB from DOL. There was temporary hack in ModifyQP
Dev Cmd to program this during Header template update operation.
To keep code simple, AdminQ will not program this area as this
CB will move out of the AH table space eventually. So, DOL will
directly write to this area after ModifyQP.

+++ nic/hal/pd/gft/rdma_pd.hpp
@@ -45,7 +45,7 @@ extern hal_ret_t p4pd_common_p4plus_txdma_stage0_rdma_params_table_entry_add (ui
                                                       uint8_t sq_qtype,
                                                       uint8_t rq_qtype,
                                                       uint8_t aq_qtype,
-                                                      uint8_t barmap_base_addr,
+                                                      uint32_t barmap_base_addr,
                                                       uint8_t barmap_size);
 extern hal_ret_t p4pd_common_p4plus_txdma_stage0_rdma_params_table_entry_get(
        uint32_t idx, tx_stage0_lif_params_table_actiondata *data);
