+++ nic/hal/pd/pd_api.hpp
@@ -2456,7 +2456,7 @@ typedef struct pd_txdma_table_entry_add_args_s {
     uint8_t sq_qtype;
     uint8_t rq_qtype;
     uint8_t aq_qtype;
-    uint64_t barmap_base_addr;
+    uint32_t barmap_base_addr;
     uint32_t barmap_size;
 } __PACK__ pd_txdma_table_entry_add_args_t;
